+++ do_test.php
@@ -96,25 +96,12 @@ function get_test_property()
  * @param string $property URL property
  * 
  * @return void
+ * 
+ * @deprecated Use do_frameset_mobile_page_content instead
  */
 function do_test_mobile_page_content($property) 
 {
-    ?>
-<div id="frame-h">
-    <iframe id="frame-h-2" 
-    src="do_test_header.php?<?php echo $property; ?>" scrolling="yes" name="h">
-    </iframe>
-</div>
-<div id="frame-ro">
-    <iframe id="frame-ro-2" src="empty.html" scrolling="yes" name="ro"></iframe>
-</div>
-<div id="frame-l">
-    <iframe  id="frame-l-2" src="empty.html" scrolling="yes" name="l"></iframe>
-</div>
-<div id="frame-ru">
-    <iframe id="frame-ru-2" src="empty.html" scrolling="yes" name="ru"></iframe>
-</div>
-        <?php 
+    do_frameset_mobile_page_content("do_test_header.php?" . $property, "empty.html", true);
 }
 
 /**
@@ -123,12 +110,14 @@ function do_test_mobile_page_content($property)
  * @param string $property URL property for HEADER
  * 
  * @return void
+ * 
+ * @deprecated Use do_test_desktop_page instead
  */
 function do_test_mobile_page($property) 
 {
     do_frameset_mobile_css();
     do_frameset_mobile_js();
-    do_test_mobile_page_content($property);
+    do_frameset_mobile_page_content("do_test_header.php?" . $property, "empty.html", true);
 }
 
 /**
@@ -186,15 +175,16 @@ onclick="hideRightFrames();">
  * Start the test page.
  * 
  * @param string $p Some property to add to the URL of do_test_test.php.
- * @param bool   $mobile Set to true to use mobile mode.
+ * 
+ * @since 2.2.1 The $mobile paramater is no longer required.
  * 
  * @return void
  */
-function do_test_page($p, $mobile)
+function do_test_page($p)
 {
     pagestart_nobody('Test');
     
-    if ($mobile && false) {
+    if (is_mobile()) {
         do_test_mobile_page($p);
     } else {
         do_test_desktop_page($p);
@@ -214,7 +204,7 @@ function do_test_page($p, $mobile)
 function try_start_test($p): void
 {
     if ($p != '') {
-        do_test_page($p, is_mobile());
+        do_test_page($p);
     } else {
         header("Location: edit_texts.php");
         exit();